modified:   lib/src/features/events/presentation/widgets/events_page.dart 	modified:   lib/src/features/news/presentation/widgets/eklase_page.dart 	modified:   lib/src/features/scanners/presentation/widgets/scanner_page.dart 	modified:   lib/src/features/scanners/presentation/widgets/scanners_page.dart 	modified:   lib/src/features/settings/application/services/settings_service.dart 	modified:   lib/src/features/settings/data/repositories/settings_repository.dart 	modified:   lib/src/features/shop/presentation/widgets/shop_page.dart 	modified:   lib/src/features/tickets/presentation/widgets/tickets_page.dart 	modified:   lib/src/features/volunteering/presentation/widgets/volunteering_jobs_page.dart

+++ lib/src/features/settings/application/services/settings_service.dart
@@ -38,7 +38,9 @@ class SettingsService {
   }
 
   Future<void> deleteUser(String avatarUrl) async {
-    await settingsRepository.deleteImage(avatarUrl, 'avatars');
+    if(avatarUrl == ""){
+      await settingsRepository.deleteImage(avatarUrl, 'avatars');
+    }
     await settingsRepository.deleteUser();
   }
 }
